requirements.txt
fuzzywuzzy==0.18.0
gTTS==2.5.3
openai==1.40.6
python-dotenv~=1.0.1
flask~=3.0.3
Markdown~=3.6
python-socketio
mistune~=3.0.2